reworked the MultiObjectiveBenchmark class
